Users/requirements.txt
flask 
pymongo
python-dotenv
pika